put tu list fixed

+++ libft/ft_strlen.c
@@ -17,9 +17,9 @@
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *str)
+int	ft_strlen(const char *str)
 {
-	size_t	i;
+	int	i;
 
 	if (str == NULL)
 		return (0);
